refactor: RepoImpl 로직(각 함수에서 서버에 같은 요청을 두 번하는 것) 수정
